Move API lookup into lookup function file
